+++ lib/pages/auth/login_page.dart
@@ -5,7 +5,7 @@ import 'package:snippet_coder_utils/ProgressHUD.dart';
 import 'package:snippet_coder_utils/hex_color.dart';
 
 import 'package:resumecraft/models/login/login_request_model.dart';
-import 'package:resumecraft/services/user_api_service.dart';
+import 'package:resumecraft/api_services/user_api_service.dart';
 import 'package:resumecraft/utils/shared_prefs/user_shared_prefs.dart';
 import 'package:resumecraft/utils/helpers/dialog_helper.dart';
