+++ lib/utils/mixins/education/education_mixin.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:resumecraft/services/education_api_service.dart';
+import 'package:resumecraft/api_services/education_api_service.dart';
 import 'package:resumecraft/utils/shared_prefs/user_shared_prefs.dart';
 import 'package:resumecraft/models/profile_section/education/read/education_model.dart';
