+++ lib/utils/mixins/experience/experiences_mixin.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'package:resumecraft/services/Experience_api_service.dart';
+import 'package:resumecraft/api_services/experience_api_service.dart';
 import 'package:resumecraft/utils/shared_prefs/user_shared_prefs.dart';
 import 'package:resumecraft/models/profile_section/experience/read/experiences_model.dart';
