+++ lib/utils/mixins/personal_detail/personal_details_mixin.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'package:resumecraft/services/personal_detail_api_service.dart';
+import 'package:resumecraft/api_services/personal_detail_api_service.dart';
 import 'package:resumecraft/utils/shared_prefs/user_shared_prefs.dart';
 import 'package:resumecraft/models/profile_section/personal_detail/read/personal_details_model.dart';
