+++ lib/utils/mixins/user/user_mixin.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:resumecraft/utils/shared_prefs/user_shared_prefs.dart';
 import 'package:resumecraft/models/profile/profile_model.dart';
-import 'package:resumecraft/services/user_api_service.dart';
+import 'package:resumecraft/api_services/user_api_service.dart';
 
 mixin UserProfileMixin<T extends StatefulWidget> on State<T> {
   String userToken = '';
